@@ -1,15 +1,15 @@
 {
-    "name": "lysice/hyperf-redis-lock",
+    "name": "pizsd/hyperf-redis-lock",
     "type": "library",
     "license": "MIT",
     "keywords": [
         "php",
         "hyperf"
     ],
-    "description": "",
+    "description": "hyperf3-redis-lock",
     "autoload": {
         "psr-4": {
-            "Lysice\\HyperfRedisLock\\": "src/"
+            "Pizsd\\HyperfRedisLock\\": "src/"
         }
     },
     "autoload-dev": {
@@ -18,16 +18,15 @@
         }
     },
     "require": {
-        "php": ">=7.3"
+        "php": ">=8.0"
     },
     "require-dev": {
-        "friendsofphp/php-cs-fixer": "^2.14",
-        "mockery/mockery": "^1.0",
-        "phpstan/phpstan": "^0.12",
-        "phpunit/phpunit": ">=7.0",
-        "swoole/ide-helper": "dev-master",
-        "swow/swow": "dev-develop",
-        "symfony/var-dumper": "^5.1"
+        "friendsofphp/php-cs-fixer": "^3.47.1",
+        "mockery/mockery": "^1.6",
+        "phpstan/phpstan": "^1.10",
+        "phpunit/phpunit": "^9.6.16",
+        "swoole/ide-helper": "^5.1.1",
+        "symfony/var-dumper": "~6.0.1"
     },
     "minimum-stability": "dev",
     "prefer-stable": true,
@@ -42,7 +41,7 @@
     },
     "extra": {
         "hyperf": {
-            "config": "Lysice\\HyperfRedisLock\\ConfigProvider"
+            "config": "Pizsd\\HyperfRedisLock\\ConfigProvider"
         }
     }
 }
