@@ -1,6 +1,8 @@
 <?php
 
-namespace Lysice\HyperfRedisLock;
+declare (strict_types=1);
+
+namespace Pizsd\HyperfRedisLock;
 
 use Hyperf\Utils\Str;
 use Hyperf\Utils\InteractsWithTime;
@@ -10,64 +12,63 @@ abstract class Lock implements LockContract
     use InteractsWithTime;
 
     /**
-     * The name of the lock
-     * @var string
-     */
-    protected $name;
-
-    /**
-     * @var int
-     */
-    protected $seconds;
-
-    /**
      * The scope identifier of this lock
+     * 
      * @var string
      */
-    protected $owner;
+    protected ?string $owner;
 
-    public function __construct($name, $seconds, $owner = null)
+    public function __construct($owner = null)
     {
-        if(is_null($owner)) {
+        if (is_null($owner)) {
             $owner = Str::random();
         }
-        $this->name = $name;
-        $this->seconds = $seconds;
         $this->owner = $owner;
     }
 
     /**
      * Attempt to acquire the lock
+     *
+     * @param string $key
+     * @param int $ttl millisecond
      * @return bool
      */
-    abstract public function acquire();
+    abstract public function acquire(string $key, int $ttl): bool;
 
     /**
      * Release the lock
-     * @return void
+     *
+     * @param string $key
+     * @return mixed
+     * @author: FatTiger{佛祖保佑代码永无BUG}
      */
-    abstract public function release();
+    abstract public function release(string $key);
 
     /**
      * Returns the owner value written into the driver for this lock
+     *
+     * @param string $key
      * @return string
      */
-    abstract protected function getCurrentOwner();
+    abstract protected function getCurrentOwner(string $key): string;
 
     /**
      * Attempt to acquire the lock
-     * @param null|\Closure $callback
-     * @param null|\Closure $finally
-     * @return bool|mixed
+     *
+     * @param string $key
+     * @param int $ttl
+     * @param $callback
+     * @param $finally
+     * @return bool
      */
-    public function get($callback = null, $finally = null)
+    public function get(string $key, int $ttl, $callback = null, $finally = null)
     {
-        $result = $this->acquire();
-        if($result && is_callable($callback)) {
+        $result = $this->acquire($key, $ttl);
+        if ($result && is_callable($callback)) {
             try {
                 return $callback();
             } finally {
-                $this->release();
+                $this->release($key);
             }
         }
         if (!$result && is_callable($finally)) {
@@ -78,26 +79,29 @@ abstract class Lock implements LockContract
     }
 
     /**
-     * @param $seconds
-     * @param null $callback
-     * @return bool|mixed
+     * Blocking lock
+     * 
+     * @param string $key
+     * @param int $ttl millisecond
+     * @param \Closure|null $callback
+     * @return mixed
      * @throws LockTimeoutException
      */
-    public function block($seconds, $callback = null)
+    public function block(string $key, int $ttl, ?\Closure $callback = null): mixed
     {
         $starting = $this->currentTime();
-        while(! $this->acquire()) {
-           usleep(250 * 1000);
-           if($this->currentTime() - $seconds >= $starting) {
-               throw new LockTimeoutException();
-           }
+        while (!$this->acquire($key, $ttl)) {
+            usleep(250 * 1000);
+            if ($this->currentTime() - ($ttl/1000) >= $starting) {
+                throw new LockTimeoutException();
+            }
         }
 
-        if(is_callable($callback)) {
+        if (is_callable($callback)) {
             try {
                 return $callback();
             } finally {
-                $this->release();
+                $this->release($key);
             }
         }
 
@@ -109,18 +113,18 @@ abstract class Lock implements LockContract
      *
      * @return string
      */
-    public function owner()
+    public function owner(): string
     {
         return $this->owner;
     }
 
     /**
      * Determines whether this lock is allowed to release the lock in the driver.
-     *
+     * @param string $key
      * @return bool
      */
-    protected function isOwnedByCurrentProcess()
+    protected function isOwnedByCurrentProcess(string $key): bool
     {
-        return $this->getCurrentOwner() === $this->owner;
+        return $this->getCurrentOwner($key) === $this->owner;
     }
 }
